cleaned up folder structures
