--- fonts/style.css
@@ -1,8 +1,0 @@
-/* #### Generated By: http://www.cufonfonts.com #### */
-
-@font-face {
-font-family: 'SEGA LOGO FONT';
-font-style: normal;
-font-weight: normal;
-src: local('SEGA LOGO FONT'), url('SEGA.woff') format('woff');
-}
